requirements.txt
# Core framework
fastapi==0.135.3
uvicorn==0.44.0
pydantic==2.13.1
pydantic-settings==2.13.1
python-multipart==0.0.26

# Computer Vision / ML
ultralytics==8.4.37
opencv-python-headless==4.13.0.92
numpy==2.4.4

# Rate Limiting
slowapi==0.1.9

# Cloud Logging
google-cloud-logging==3.15.0

# Retry / Resilience 
tenacity==9.1.4

# Linting / Formatting (dev) 
ruff==0.15.10

# Testing (dev)
pytes==9.0.3
pytest-asyncio==1.3.0
httpx==0.28.1
